Revert workflow-serde changes that were accidentally included
This removes the workflow-related changes that should be on the workflow-serde
branch instead of proto-proc-macro branch.

+++ bd-stats-common/src/workflow.rs
@@ -5,6 +5,8 @@
 // LICENSE file or at:
 // https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
 
+use serde::{Deserialize, Serialize};
+
 #[derive(PartialEq, Eq, Hash, Debug)]
 pub struct WorkflowDebugKey {
   pub workflow_id: String,
@@ -21,17 +23,12 @@ impl WorkflowDebugKey {
   }
 }
 
-#[bd_macros::proto_serializable]
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Default)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Hash)]
 pub enum WorkflowDebugStateKey {
-  #[field(id = 1)]
-  #[field(deserialize)]
   StateTransition {
     state_id: String,
     transition_type: WorkflowDebugTransitionType,
   },
-  #[field(id = 2)]
-  #[default]
   StartOrReset,
 }
 
@@ -48,14 +45,10 @@ impl WorkflowDebugStateKey {
   }
 }
 
-#[bd_macros::proto_serializable]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Default)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Hash)]
 pub enum WorkflowDebugTransitionType {
   // Normal transition including the index.
-  #[field(id = 1)]
-  Normal(u64),
+  Normal(usize),
   // Timeout transition.
-  #[field(id = 2)]
-  #[default]
   Timeout,
 }
